pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plc_vs"
version = "0.1.0"
description = "Agentic system for answering questions about PLC projects"
readme = "docs/README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pydantic-ai[openai]>=0.0.14",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ingestion*", "agents*", "tools*"]
